Move CoreDump into test/e2e/util.go, add docstring.
